store/tikv:move TxnScope into tikv config
